+++ packages/perfrunner-cli/src/cli.ts
@@ -1,20 +1,16 @@
 #!/usr/bin/env node
 
 import { logger } from 'perfrunner-core';
-
-import { createCommand } from './commands';
-import { parseConsole } from './console/console-arguments-parser';
+import { parseUserInput, getCommandName } from './console/parser';
 import { setupLogLevel } from './logger';
+import { getCommandFromUserInput } from './commands';
 
 (async function (): Promise<number> {
     try {
-        const userInput = parseConsole();
-
-        setupLogLevel(userInput.logLevel);
-
-        const commandName = userInput._unknown == null ? '--from-console' : userInput._unknown[0];
-        const cmd = createCommand(commandName, userInput);
-
+        const inputParams = parseUserInput();
+        setupLogLevel(inputParams.logLevel);
+        const commandName = getCommandName(inputParams);
+        const cmd = getCommandFromUserInput(commandName, inputParams);
         return await cmd.execute();
     } catch (error) {
         logger.error(error);
