+++ packages/perfrunner-cli/src/commands/index.ts
@@ -1,6 +1,3 @@
-import { RunTestsFromConfigCommand } from './from-config-command';
-import { RunTestsFromConsoleCommand } from './from-console-command';
-import { InitConfigCommand } from './init-command';
-import { createCommand } from './command-factory';
+import { getCommandFromUserInput } from './command-factory';
 
-export { RunTestsFromConfigCommand, RunTestsFromConsoleCommand, InitConfigCommand, createCommand };
+export { getCommandFromUserInput };
