+++ packages/perfrunner-cli/src/console/custom-types/index.ts
@@ -1,8 +1,8 @@
-import { Network, HSPA } from './network';
+import { Network } from './network';
 import { Url } from './url';
 import { ArgsLikeString } from './arg-like-string';
 import { StringOrNumber } from './string-number';
 import { LogLevel } from './log-level';
 import { Bool } from './bool';
 
-export { Network, Url, ArgsLikeString, StringOrNumber, LogLevel, HSPA, Bool };
+export { Network, Url, ArgsLikeString, StringOrNumber, LogLevel, Bool };
