--- packages/perfrunner-cli/src/reporter/index.ts
@@ -1,3 +1,0 @@
-import { loadReporter } from './loader';
-
-export { loadReporter };
