+++ packages/perfrunner-cli/src/runner/generate-report.ts
@@ -1,5 +1,5 @@
 import { logger } from 'perfrunner-core';
-import { loadReporter } from '../reporter';
+import { loadReporter } from './report-loader';
 
 export async function generateReport(name: string, outputTo: string, data: any, args: string[]) {
     const reporter = await loadReporter(name);
